automation_foundation/requirements.txt

fastapi==0.112.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
redis==5.0.8
rq==1.16.2
Jinja2==3.1.4
WeasyPrint==61.2
pdfplumber==0.11.4
playwright==1.46.0
requests==2.32.3
python-slugify==8.0.4
